Remove unused webhooks and combine to single file
