--- netlify/functions/comments.js
@@ -1,4 +1,0 @@
-export async function addComment(context, message, owner, repo, number) {
-    const issueComment = { owner: owner, repo: repo, issue_number: number, body: message };
-    await context.octokit.issues.createComment(issueComment);
-}
